+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 [unreleased]
+* add `wp_get_runtime_environment()` to return value of constant `WP_RUNTIME_ENVIRONMENT`
+
 #### 1.3.2 / 2022-02-15
 * update to correspond to core patch